+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Experimental disclaimer warning when creating an x402 wallet via `mcpc x402 init`
+
+- Experimental feature warning for `x402 init` and `x402 import` commands
 
 ## [0.2.3] - 2026-03-31